+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Minor Changes
 - 1833 | Country in error for Cloud Service and DC
 - 1876 | fix the header for digital service detail page / for inventory - Equipment and Application view / for ecomindIA / for administration
+- 1861 | Digital service : Add raw value into camembert
 
 ## 3.9.1